Home Page
I set up a basic homepage which is linked to the style sheet named "style" as well as 2 textboxes for guessing the song and artist in php. Unfortunately the "submit" button appears above the textboxes for some reason--I'll figure out how to fix it. The textboxes connect to a "processGuess" php file. There is also a button to play more time into the song which is a javascript file.